Move some internals into `Pipeline` folder
