+++ src/Polly.Core/Utils/Pipeline/BridgeComponent.TResult.cs
@@ -1,0 +1,37 @@
+﻿namespace Polly.Utils.Pipeline;
+
+[DebuggerDisplay("{Strategy}")]
+internal sealed class BridgeComponent<T> : BridgeComponentBase
+{
+    public BridgeComponent(ResilienceStrategy<T> strategy)
+        : base(strategy) => Strategy = strategy;
+
+    public ResilienceStrategy<T> Strategy { get; }
+
+    internal override ValueTask<Outcome<TResult>> ExecuteCore<TResult, TState>(
+        Func<ResilienceContext, TState, ValueTask<Outcome<TResult>>> callback,
+        ResilienceContext context,
+        TState state)
+    {
+        // Check if we can cast directly, thus saving some cycles and improving the performance
+        if (callback is Func<ResilienceContext, TState, ValueTask<Outcome<T>>> casted)
+        {
+            return TaskHelper.ConvertValueTask<T, TResult>(
+                Strategy.ExecuteCore(casted, context, state),
+                context);
+        }
+        else
+        {
+            var valueTask = Strategy.ExecuteCore(
+                static async (context, state) =>
+                {
+                    var outcome = await state.callback(context, state.state).ConfigureAwait(context.ContinueOnCapturedContext);
+                    return outcome.AsOutcome<T>();
+                },
+                context,
+                (callback, state));
+
+            return TaskHelper.ConvertValueTask<T, TResult>(valueTask, context);
+        }
+    }
+}
