+++ test/Polly.Core.Tests/ResiliencePipelineTTests.Sync.cs
@@ -1,4 +1,5 @@
 using Polly.Utils;
+using Polly.Utils.Pipeline;
 
 namespace Polly.Core.Tests;
 
@@ -79,7 +80,7 @@ public partial class ResiliencePipelineTests
     [Theory]
     public void Execute_GenericStrategy_Ok(Action<ResiliencePipeline<string>> execute)
     {
-        var pipeline = new ResiliencePipeline<string>(PipelineComponent.FromStrategy(new TestResilienceStrategy
+        var pipeline = new ResiliencePipeline<string>(PipelineComponentFactory.FromStrategy(new TestResilienceStrategy
         {
             Before = (c, _) =>
             {
